requirements.txt
tensorflow-gpu==1.10.0
scikit-learn==0.19.1
numpy==1.15.4
pandas==0.23.4
jieba==0.39
